Redesigned logic structure of code for "view" block.
