+++ src/com/leofesk/quicktodomanager/view/notes/AddFrame.java
@@ -1,30 +1,31 @@
-package com.leofesk.quicktodomanager.view;
+package com.leofesk.quicktodomanager.view.notes;
 
 import com.leofesk.quicktodomanager.controller.DataBaseWorker;
+import com.leofesk.quicktodomanager.view.MainFrame;
 
 import javax.swing.*;
 import java.awt.*;
 
-public class NoteEditFrame extends JFrame {
-    private JButton buttonAddOrEdit;
+public class AddFrame extends JFrame {
+    private JButton buttonAdd;
     private JLabel labelTaskName;
     private JLabel labelDeadline;
     private JLabel labelFormatDeadline;
     private JScrollPane scrollPaneForTextArea;
-    private static JTextArea textArea;
-    private static JTextField textFieldDeadlineDate;
-    private static JTextField textFieldTaskName;
+    private JTextArea textArea;
+    private JTextField textFieldDeadlineDate;
+    private JTextField textFieldTaskName;
     private static ImageIcon imageAppIcon;
     private String pathToAppLogo = "/img/AppLogo.png";
 
-    public NoteEditFrame() {
+    public AddFrame() {
         initComponents();
     }
 
     private void initComponents() {
         scrollPaneForTextArea = new JScrollPane();
         textArea = new JTextArea();
-        buttonAddOrEdit = new JButton();
+        buttonAdd = new JButton();
         labelTaskName = new JLabel();
         textFieldDeadlineDate = new JTextField();
         labelDeadline = new JLabel();
@@ -33,26 +34,24 @@ public class NoteEditFrame extends JFrame {
         imageAppIcon = new ImageIcon(MainFrame.class.getResource(pathToAppLogo));
         setIconImage(imageAppIcon.getImage());
 
-        setTitle(" Edit note");
+        setTitle(" Add new note");
         setDefaultCloseOperation(WindowConstants.DISPOSE_ON_CLOSE);
         setResizable(false);
 
         textArea.setColumns(20);
         textArea.setRows(5);
-        textArea.setWrapStyleWord(true);
         textArea.setLineWrap(true);
         textArea.setMargin(new Insets(5, 5, 5, 5));
         scrollPaneForTextArea.setViewportView(textArea);
 
-        buttonAddOrEdit.addActionListener(e -> actionButtonEdit());
-        buttonAddOrEdit.setText("ОК");
+        buttonAdd.addActionListener(e -> actionButtonAdd());
+        buttonAdd.setText("ОК");
 
         labelTaskName.setText("Title:");
         textFieldDeadlineDate.setText("11.08.1992");
         labelDeadline.setText("Deadline:");
         labelFormatDeadline.setText("Format: DD.MM.YYYY");
-        textFieldTaskName.setText("Name Task");
-
+        textFieldTaskName.setText("");
 
         GroupLayout layout = new GroupLayout(getContentPane());
         getContentPane().setLayout(layout);
@@ -73,11 +72,11 @@ public class NoteEditFrame extends JFrame {
                                                         .addGroup(layout.createSequentialGroup()
                                                                 .addComponent(labelDeadline)
                                                                 .addGap(10, 10, 10)
-                                                                .addComponent(textFieldDeadlineDate, GroupLayout.PREFERRED_SIZE, 84, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                                                .addComponent(textFieldDeadlineDate, GroupLayout.PREFERRED_SIZE, 84, GroupLayout.PREFERRED_SIZE)
                                                                 .addGap(10, 10, 10)
                                                                 .addComponent(labelFormatDeadline, GroupLayout.DEFAULT_SIZE, 198, Short.MAX_VALUE)
                                                                 .addGap(10, 10, 10)
-                                                                .addComponent(buttonAddOrEdit, GroupLayout.PREFERRED_SIZE, 131, javax.swing.GroupLayout.PREFERRED_SIZE)))))
+                                                                .addComponent(buttonAdd, GroupLayout.PREFERRED_SIZE, 131, GroupLayout.PREFERRED_SIZE)))))
                                 .addGap(10, 10, 10))
         );
 
@@ -94,7 +93,7 @@ public class NoteEditFrame extends JFrame {
                                 .addGroup(layout.createParallelGroup(GroupLayout.Alignment.BASELINE)
                                         .addComponent(labelDeadline, GroupLayout.PREFERRED_SIZE, 25, GroupLayout.PREFERRED_SIZE)
                                         .addComponent(textFieldDeadlineDate, GroupLayout.PREFERRED_SIZE, 25, GroupLayout.PREFERRED_SIZE)
-                                        .addComponent(buttonAddOrEdit)
+                                        .addComponent(buttonAdd)
                                         .addComponent(labelFormatDeadline, GroupLayout.PREFERRED_SIZE, 24, GroupLayout.PREFERRED_SIZE))
                                 .addGap(10, 10, 10))
         );
@@ -103,24 +102,17 @@ public class NoteEditFrame extends JFrame {
         setLocationRelativeTo(null);
     }
 
-    private void actionButtonEdit() {
+    private void actionButtonAdd() {
         if (DataBaseWorker.isCorrectNote(textFieldTaskName.getText(), textArea.getText(), textFieldDeadlineDate.getText())) {
-            DataBaseWorker.editNoteFromTable(textFieldTaskName.getText(), textArea.getText(), textFieldDeadlineDate.getText());
+            DataBaseWorker.addNewNoteFromTable(textFieldTaskName.getText(), textArea.getText(), textFieldDeadlineDate.getText());
         } else {
-            DataBaseWorker.showMessage("Task was not updated. Title, text and deadline can't be empty. [CODE:V_NE_001]");
+            DataBaseWorker.showMessage("New task was not created. Title, text and deadline can't be empty.");
         }
+
+        textFieldTaskName.setText("");
+        textArea.setText("");
+        textFieldDeadlineDate.setText("01.01.2000");
+
         dispose();
-    }
-
-    public static void setTextArea(String text) {
-        textArea.setText(text);
-    }
-
-    public static void setTextFieldDeadlineDate(String deadline) {
-        textFieldDeadlineDate.setText(deadline);
-    }
-
-    public static void setTextFieldTaskName(String title) {
-        textFieldTaskName.setText(title);
     }
 }
